re-orient interactive linters
Instead of running every linter on every entity, we run every entity
through every linter. Found it kind of confusing to be in interactive
mode and to have to keep switching between linters mentally. Still has
some flaws in that it bumps up and down the tree, so it'll do all of the
tasks for one list, then do the list, then back down to more tasks. More
to the point of having a flat structure instead of a tree.