Detect hashtag when checking for context link
**what**
Some usability improvements to InteractiveLinter::MissingContext.

- When adding context links, if it can find a project card with a
matching hashtag, will ask for y/n confirmation to add a context link
for that project.

- If it can't find a match, it'll default to previous
behavior - list all project cards and take a user-inputted number to
pick the right one.

- updates options so that you can enter 'n' if you dont want to take the
suggestion and are given a list of all projects to choose from

- general updates so the whole thing doesn't quit when you mess up, but
just moves on to the next thing and tells you it skipped that card.

- TESTS!!!!