Infer pound symbol when assigning hashtags
**What**

Instead of requiring the user to type the pound symbol (`#`) every time
they want to add a hashtag to a Trello card, this PR will infer the
hashtag.

Also adds a test file for the whole shebang.

**Why**

Hopefully this will make things less tedious. Additionally, based on the
language in the message, it was unclear as to whether or not you needed to
put the pound symbol when assigning hashtags.