Change `raise` to `warn` on invalid configuration
This becomes a problem when you add keys to the config file for new
features that are not yet merged into master. Right now we have a branch
that makes use of `github_reviewers` inside of `.scrum-lint.yml`, but
then the build fails on master because it doesn't support that
configuration. A warning seems more appropriate for this case.